# docs/openapi.yaml
openapi: 3.0.0
info:
  title: EventHorizon Gateway API
  description: API для фронтенда (React/Angular)
  version: 1.0.3
servers:
  - url: http://localhost:8079
    description: Local development (через балансировщик)
  - url: http://localhost:8081
    description: Gateway-1 (напрямую)
  - url: http://localhost:8082
    description: Gateway-2 (напрямую)
  - url: http://localhost:8083
    description: Gateway-3 (напрямую)
paths:
...
  /api/auth/register:
    post:
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 6
                  example: secret123
                nickname:
                  type: string
                  description: Опционально. Если не указан, генерируется из email
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  success:
                    type: boolean
                  message:
                    type: string
  /api/auth/login:
    post:
      summary: Login user (получить JWT токен)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: user@example.com
                password:
                  type: string
                  example: secret123
      responses:
        '200':
          description: Success
          content:
            application/json:
            schema:
              type: object
              properties:
                access_token:
                  type: string
                  description: JWT токен (Bearer)
                token_type:
                  type: string
                  example: Bearer
                expires_in:
                  type: integer
                  example: 86400
                user_id:
                  type: string
  /api/auth/user:
    get:
      summary: Get user data (email, nickname)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  nickname:
                    type: string
                  best_scores:
                    type: object
                    additionalProperties:
                      type: integer
                  total_score:
                    type: integer
        '401':
          description: Unauthorized
  /api/auth/update-nickname:
    post:
      summary: Update user nickname
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nickname
              properties:
                nickname:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '401':
          description: Unauthorized

  # ===== BILLING =====
  /api/billing/balance/all:
    get:
      summary: Get all balances (lamps + tickets)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lamps:
                    type: integer
                  tickets:
                    type: integer
        '401':
          description: Unauthorized

  # ===== GAME =====
  /api/game/submit:
    post:
      summary: Submit game score (server-side validation)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - game_id
                - score
                - seed
              properties:
                user_id:
                  type: string
                game_id:
                  type: string
                  enum: [hexagon, flappy, memory, towers]
                level:
                  type: integer
                  default: 1
                score:
                  type: integer
                  description: Количество очков (начисленных игрой)
                user_email:
                  type: string
                nickname:
                  type: string
                seed:
                  type: string
                  description: Deterministic seed для валидации
                moves:
                  type: array
                  items:
                    type: object
                    properties:
                      fromX:
                        type: integer
                      fromY:
                        type: integer
                      toX:
                        type: integer
                      toY:
                        type: integer
                      timestamp:
                        type: integer
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  new_highscore:
                    type: integer
                  message:
                    type: string
                  lamps_earned:
                    type: integer
                  tickets_earned:
                    type: integer
        '401':
          description: Unauthorized

  # ===== LEADERBOARD =====
  /api/leaderboard:
    get:
      summary: Get top 10 players
      parameters:
        - name: game_id
          in: query
          required: true
          schema:
            type: string
            enum: [hexagon, flappy, memory, towers]
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            maximum: 100
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  entries:
                    type: array
                    items:
                      type: object
                      properties:
                        rank:
                          type: integer
                        user_id:
                          type: string
                        user_email:
                          type: string
                        score:
                          type: integer
                        updated_at:
                          type: integer
                          description: Unix timestamp

  # ===== PROFILE =====
  /api/profile:
    get:
      summary: Get aggregated user profile (CQRS read model)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                  email:
                    type: string
                  nickname:
                    type: string
                  total_score:
                    type: integer
                  best_scores:
                    type: object
                    additionalProperties:
                      type: integer
                  lamps:
                    type: integer
                  tickets:
                    type: integer
        '401':
          description: Unauthorized
        '404':
          description: Profile not found

  # ===== SHOP =====
  /api/shop/items:
    get:
      summary: Get shop items (with filters)
      security:
        - BearerAuth: []
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [all, game_skin, merch, profile_theme]
          description: Filter by category
        - name: game_id
          in: query
          schema:
            type: string
            enum: [hexagon, flappy, memory, towers]
          description: Filter by game (for skins)
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    name:
                      type: string
                    description:
                      type: string
                    price:
                      type: integer
                    category:
                      type: string
                    game_id:
                      type: string
                    image_url:
                      type: string
                    available:
                      type: boolean
                    owned:
                      type: boolean
        '401':
          description: Unauthorized
  /api/shop/purchase:
    post:
      summary: Purchase an item
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - item_id
              properties:
                item_id:
                  type: string
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  new_balance:
                    type: integer
        '400':
          description: Bad request (not enough tickets, already owned, etc.)
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT